+++ packages/opds-web-client/CHANGELOG.md
@@ -1,8 +1,16 @@
 ## Changelog
 
-### v0.4.6
+### v0.5.1
 
 - Fix: Use `new URL()` instead of `url.resolve` when parsing OPDS links. `resolve` is legacy and was causing bugs when on `https` but trying to resolve a link with a nested `http` segment.
+
+### v0.4.7
+
+- Display "Read online" instead of "Download..." in download buttons for AxisNow media type
+
+### v0.4.6
+
+- Add AxisNow media type
 
 ### v0.4.5